connect fields to superdoc + minor refactoring
