+++ packages/super-editor/src/index.js
@@ -4,11 +4,12 @@ import { makeDefaultItems } from '@/components/toolbar/defaultItems.js';
 import { getMarksFromSelection, getActiveFormatting } from '@/core/helpers/getMarksFromSelection.js';
 import { getStarterExtensions } from '@extensions/index.js';
 import { SuperToolbar } from '@components/toolbar/super-toolbar.js';
-import DocxZipper from '@core/DocxZipper';
+import { DocxZipper, helpers } from '@core/index.js';
 import SuperEditor from '@components/SuperEditor.vue';
 import BasicUpload from './dev/components/BasicUpload.vue';
 import Toolbar from '@components/toolbar/Toolbar.vue';
 import SuperInput from '@components/SuperInput.vue';
+import * as fieldAnnotationHelpers from '@extensions/field-annotation/fieldAnnotationHelpers/index.js';
 
 /**
  * Exported classes and components.
@@ -31,9 +32,10 @@ export {
   Toolbar,
 
   // Helpers
+  helpers,
+  fieldAnnotationHelpers,
   makeDefaultItems,
   getMarksFromSelection,
   getActiveFormatting,
   getStarterExtensions,
-
 }
